% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountyLevelPollutantValue.R
\name{getCountyLevelPollutantValue}
\alias{getCountyLevelPollutantValue}
\title{Get county level pollutant values for a year}
\usage{
getCountyLevelPollutantValue(year, pollutant = "PM2.5", counties_shp)
}
\arguments{
\item{year}{numeric represented as yyyy. Earliest available year: "1997".}

\item{pollutant}{string, can be one of "PM2.5", "Ozone", "NO2", "SO2" or "CO". Default set to "PM2.5"}

\item{counties_shp}{dataframe extracted from tigris library that contains the tract names and geometry of all census tracts (tracts(state = NULL, county = NULL, cb = TRUE, year = NULL))}
}
\description{
Given a year and type of pollutant, the function will return the corresponding mean pollutant values for each county.
}
