% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMonthPollutionEstimate.R
\name{getMonthPollutionEstimate}
\alias{getMonthPollutionEstimate}
\title{Get pollution estimate for a given month}
\usage{
getMonthPollutionEstimate(long, lat, pollutant = "PM2.5", monthyear)
}
\arguments{
\item{long}{numeric longitude value}

\item{lat}{numeric latitude value}

\item{pollutant}{string, one of "PM2.5", "Ozone", "NO2", "SO2" or "CO". Default set to "PM2.5"}

\item{monthyear}{string represented as "mm-yyyy". Example: "12-2015".}
}
\description{
Get estimate of a given pollutant for a given month ("01-2005" - "12-2017"). Derived from daily EPA files.
}
\examples{
getMonthPollutionEstimate(-75.162, 39.9526, "PM2.5", "12-2015")
}
