% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPollutionEstimates.df.R
\name{getPollutionEstimates.df}
\alias{getPollutionEstimates.df}
\title{Get pollution estimates for dataset}
\usage{
getPollutionEstimates.df(data, monthyear_start, monthyear_end)
}
\arguments{
\item{data}{dataframe with numeric Longitude and Latitude columns}

\item{monthyear_start}{string represented as "month-year". Earliest available month-year: "01-2005".}

\item{monthyear_end}{string represented as "month-year". Latest available month-year: "12-2017".}
}
\description{
Given a dataset with Longitude and Latitude columns, will return dataset with columns for each pollutant corresponding to the average of monthly estimates spanning requested time period. Monthly estimates derived from daily EPA files.
}
\examples{
Longitude <- c(-75.133346, -96.27017, -80.448374)
Latitude <- c(40.009376, 29.891901, 26.649124)
dat <- data.frame(Longitude, Latitude)
getPollutionEstimates.df(dat, "01-2005", "12-2017")
}
