% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRangePollutionEstimate.R
\name{getRangePollutionEstimate}
\alias{getRangePollutionEstimate}
\title{Get pollution estimate over time period}
\usage{
getRangePollutionEstimate(long, lat, pollutant = "PM2.5", monthyear_start,
  monthyear_end, result = "mean")
}
\arguments{
\item{long}{numeric longitude value}

\item{lat}{numeric latitude value}

\item{pollutant}{string, one of "PM2.5", "Ozone", "NO2", "SO2" or "CO". Default set to "PM2.5"}

\item{monthyear_start}{string represented as "mm-yyyy." Earliest available month-year: "01-2005".}

\item{monthyear_end}{string represented as "mm-yyyy." Latest available month-year: "12-2017".}

\item{result}{string, one of "mean" or "array." Mean will return the average of the monthly estimates over the time
period, array will return a vector where each value corresponds to each month over the time period}
}
\description{
Get estimate of a given pollutant for a given time period as defined by
month-year to month-year. The result can be an average over the time period or an array with
values corresponding to each month. Derived from daily EPA files.
}
\examples{
getRangePollutionEstimate(-75.162, 39.9526, "PM2.5", "01-2005", "12-2017")
}
