% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRasterPR.R
\name{getRasterPR}
\alias{getRasterPR}
\title{Get raster layer for a given year or month-year for Puerto Rico.}
\usage{
getRasterPR(pollutant, month_or_year)
}
\arguments{
\item{pollutant}{string, one of "PM2.5", "Ozone", "NO2", "SO2", "CO"}

\item{month_or_year}{string represented as "mm-yyyy" or "yyyy". Ex: "02-2012" or "2012".}
}
\description{
Get pre-generated raster layer for a given pollutant for a month or year in 2005 to 2017.
}
\examples{
getYearPollutionEstimate("PM2.5", "02-2012")
}
