% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYearPollutionEstimate.R
\name{getYearPollutionEstimate}
\alias{getYearPollutionEstimate}
\title{Get pollution estimate for a given year}
\usage{
getYearPollutionEstimate(long, lat, pollutant = "PM2.5", year)
}
\arguments{
\item{long}{numeric longitude value}

\item{lat}{numeric latitude value}

\item{pollutant}{string, one of "PM2.5", "Ozone", "NO2", "SO2" or "CO". Default set to "PM2.5"}

\item{year}{string represented as "year." Example: "2015".}
}
\description{
Get estimate of a given pollutant for a given year (2005 to 2017). Derived from annual EPA files.
}
\examples{
getYearPollutionEstimate(-75.162, 39.9526, "PM2.5", "2015")
}
