% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria_pollutants.R
\name{list_pollutant_standards}
\alias{list_pollutant_standards}
\title{Look up National Ambient Air Quality Standards (NAAQS)}
\usage{
list_pollutant_standards(parameter_code = NULL, detail = FALSE)
}
\arguments{
\item{parameter_code}{An AQS parameter code to retrieve specific standard
information. If \code{NULL} (default), display all pollutant standards.}

\item{detail}{A logical value indicating whether detailed information is
retrieved.}
}
\value{
A data.frame containing the pollutant standard information
}
\description{
A function to show pollutant standard information on criteria air
pollutants.
}
\examples{
list_pollutant_standards()

}
\references{
Data source:
\url{https://aqs.epa.gov/aqsweb/documents/codetables/pollutant_standards.html}
}
