% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozone20km.R
\docType{data}
\name{ozone20km}
\alias{ozone20km}
\title{Ozone concentration data}
\format{
A stars object with 4 dimensions and 1 attribute
\describe{
\item{x, y}{dimension; raster coordinates}
\item{event}{dimension; exceptional event filter}
\item{year}{dimension; years of data available}
\item{ozone_8_hour_2015}{attribute; ozone levels based on Ozone 8-hour 2015
standard}
}
}
\usage{
ozone20km
}
\description{
A stars object containing ozone concentrations covering the conterminous
United States for years 2021-2022 on a 20km x 20km grid.
}
\keyword{datasets}
