% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pargasite.R
\name{run_pargasite}
\alias{run_pargasite}
\title{Run PARGASITE application}
\usage{
run_pargasite(
  x,
  summarize_by = c("state", "county", "cbsa"),
  summary_state = NULL,
  summary_county = NULL,
  summary_cbsa = NULL
)
}
\arguments{
\item{x}{A stars object created by \link{create_pargasite_data}.}

\item{summarize_by}{A vector of strings specifying geographic boundaries to
compute areal means.}

\item{summary_state}{A stars object created by \link{summarize_by_boundaries}.
Ignored if 'state' is not specified in \code{summarize_by}. While the function
computes areal summary if the argument is \code{NULL}, it may be useful if the
app is to be called multiple times to save processing time.}

\item{summary_county}{A stars object created by \link{summarize_by_boundaries}.
Ignored if 'county' is not specified in \code{summarize_by}.}

\item{summary_cbsa}{A stars object created by \link{summarize_by_boundaries}.
Ignored if 'cbsa' is not specified in \code{summarize_by}.}
}
\value{
This function normally does not return; interrupt R to stop the
application (usually by pressing Ctrl + C or ESC).
}
\description{
Launch a Shiny application to visualize pollutant levels of the conterminous
US. The system's default web browser will be launched automatically after
the app is started.
}
\examples{

if (interactive()) {
  run_pargasite(ozone20km)
}

}
