% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_by.R
\name{summarize_by_boundaries}
\alias{summarize_by_boundaries}
\title{Summarized PARGASITE pollutant data by geographic boundaries}
\usage{
summarize_by_boundaries(x, level = c("state", "county", "cbsa"))
}
\arguments{
\item{x}{A stars object created by \link{create_pargasite_data}.}

\item{level}{A string specifying a geographic boundary.}
}
\value{
A stars object containing pollutant levels summarized by geographic
boundaries.
}
\description{
A function to compute areal means of grid samples falling inside the target
geographic boundaries, including State, County, and Core Based Statistical
Area (CBSA). The US shape files are automatically downloaded to R's
temporary directory once for the current session and removed when R is
closed.
}
\examples{

## State-level summary
summarize_by_boundaries(ozone20km, "state")

}
