% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pargasite-package.R
\docType{package}
\name{pargasite-package}
\alias{pargasite}
\alias{pargasite-package}
\title{pargasite: Pollution-Associated Risk Geospatial Analysis Site}
\description{
Offers tools to estimate and visualize levels of major pollutants (CO, NO2, SO2, Ozone, PM2.5 and PM10) across the conterminous United States for user-defined time ranges. Provides functions to retrieve pollutant data from the U.S. Environmental Protection Agency’s 'Air Quality System' (AQS) API service \url{https://aqs.epa.gov/aqsweb/documents/data_api.html} for interactive visualization through a 'shiny' application, allowing users to explore pollutant levels for a given location over time relative to the National Ambient Air Quality Standards (NAAQS).
}
\author{
\strong{Maintainer}: Jaehyun Joo \email{jaehyunjoo@outlook.com}

Authors:
\itemize{
  \item Rebecca Greenblatt
  \item Avantika Diwadkar
  \item Nisha Narayanan
  \item Blanca Himes
}

}
