% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pargasite-package.R
\docType{package}
\name{pargasite-package}
\alias{pargasite}
\alias{pargasite-package}
\title{pargasite: Pollution-Associated Risk Geospatial Analysis}
\description{
This package provides raster layers from which pollution estimates can be extracted for latitude longitude coordinates within the contiguous United States at a 10km x 10km resolution. The recent modified version of the package also allows the user to download the pollutant data at the county, CBSA, census tract and zipcode level. Data is based on publicly available, pre-generated files from the United States Environmental Protection Agency (EPA) for years 1997 to 2021.
}
\author{
\strong{Maintainer}: Jaehyun Joo \email{jaehyun.joo@outlook.com}

Authors:
\itemize{
  \item Rebecca Greenblatt
  \item Avantika Diwadkar
  \item Nisha Narayanan
  \item Blanca Himes
}

}
