% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pargasite-package.R
\docType{package}
\name{pargasite-package}
\alias{pargasite}
\alias{pargasite-package}
\title{pargasite: Pollution-Associated Risk Geospatial Analysis}
\description{
Pollution-Associated Risk Geospatial Analysis STIE (PARGASITE) offers tools to estimate and visualize major pollutant levels (CO, NO2, SO2, Ozone, PM2.5 and PM10) covering the conterminous United States at user-defined time ranges. It help users to automatically retrieves pollutant data via the Environmental Protection Agency’s (EPA) Air Quality System (AQS) API service, filters the data by exceptional event (e.g., wildfire) status, performs spatial interpolations, and summarizes pollutant concentrations by geographic boundaries including State, County, and Core-Based Statistical Area (CBSA). It also offers a Shiny application to interactively visualize pollutant levels and to easily check the air quality status of a given location over time relative to the level of the National Ambient Air Quality Standards (NAAQS).
}
\author{
\strong{Maintainer}: Jaehyun Joo \email{jaehyun.joo@outlook.com}

Authors:
\itemize{
  \item Rebecca Greenblatt
  \item Avantika Diwadkar
  \item Nisha Narayanan
  \item Blanca Himes
}

}
