% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pargasite.R
\name{run_pargasite}
\alias{run_pargasite}
\title{Run pargasite application}
\usage{
run_pargasite(x)
}
\arguments{
\item{x}{A stars object returned by \link{create_pargasite_data}.}
}
\value{
This function normally does not return; interrupt R to stop the
application (usually by pressing Ctrl + C or ESC).
}
\description{
Launch a Shiny application to visualize pollutant levels of the conterminous
US. The system's default web browser will be launched automatically after
the app is started.
}
\examples{

if (interactive()) {
  run_pargasite(ozone20km)
}

}
